requirements.txt
discord.py>=2.0.0
customtkinter==5.2.1
pillow==10.0.0
requests==2.31.0